+++ api/activity_picture/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from rest_framework.fields import JSONField
 
 from ..study.models import Study
