+++ api/activity_picture/serializers.py
@@ -1,7 +1,6 @@
 from drf_yasg import openapi
 from rest_framework import serializers
 from .models import ActivityPicture
-# from ..study.serializers import StudySerializer
 
 
 class ActivityPictureSerializer(serializers.ModelSerializer):
