+++ api/admin.py
@@ -5,7 +5,6 @@ from .user.models import User
 from .study_member.models import StudyMember
 from .activity_picture.models import ActivityPicture
 from .schedule.models import Schedule
-from .user_image_upload.models import TestImage
 
 admin.site.register(Category)
 admin.site.register(User)
@@ -13,4 +12,3 @@ admin.site.register(Study)
 admin.site.register(StudyMember)
 admin.site.register(ActivityPicture)
 admin.site.register(Schedule)
-admin.site.register(TestImage)
