+++ api/schedule/serializers/schedule_delete_sz.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import Schedule
+
+
+class ScheduleDeleteSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Schedule
+        fields = ['study', 'title', ]
