+++ api/study/serializers/schedule_of_study_sz.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+from ..models import Study
+from ...schedule.serializers.schedule_sz import ScheduleSerializer
+
+
+class ScheduleOfStudySerializer(serializers.ModelSerializer):
+    study_schedule = ScheduleSerializer(source='schedule_set', many=True)
+
+    class Meta:
+        model = Study
+        fields = ['study_schedule', ]
