+++ api/study/serializers/study_add_sz.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from ..models import Study
+
+class StudyAddSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Study
+        fields = ['study_id', 'category', 'title', 'limit', 'description','study_image', ]
