+++ api/study/serializers/study_sz.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import Study
+
+
+class StudySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Study
+        fields = ['study_id', 'category', 'title', 'limit', 'description', ]
