+++ api/study_member/serializers/study_add_study_member_sz.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from ..models import StudyMember
+
+
+class StudyAddStudyMemberSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = StudyMember
+        fields = ['study', 'user', 'is_manager' ]
+
+
