+++ api/study_member/serializers/study_member_delete_sz.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import StudyMember
+
+
+class StudyMemberDeleteSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = StudyMember
+        fields = ['study', 'user', ]
