--- api/user_image_upload/models.py
@@ -1,6 +1,0 @@
-from django.db import models
-
-
-# 사용자
-class TestImage(models.Model):
-    s3_profile_img = models.FileField(blank=True, null=True)
