--- api/user_image_upload/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from .models import TestImage
-
-class TestImageSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestImage
-        fields = '__all__'
